+++ lib/components/Insights/Types/Area/index.stories.tsx
