+++ lib/components/Insights/Types/Area/index.tsx
