+++ lib/components/Insights/Types/utils/colors.tsx
